Improves alerting stages across the app
